Added start node type and did some refactoring of folder structure
